Refactor out viewMapAs
